Finalize command refactoring
