+++ src/main/java/net/porillo/objects/GPlayer.java
@@ -3,6 +3,8 @@ package net.porillo.objects;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
 
 import java.util.UUID;
 
@@ -29,4 +31,14 @@ public class GPlayer {
 	 * Numerical "carbon score" value for just this player
 	 */
 	private int carbonScore;
+
+
+	public Player getPlayer() {
+		return Bukkit.getPlayer(uuid);
+	}
+
+	public void sendMsg(String msg) {
+		getPlayer().sendMessage(msg);
+	}
+
 }
